+++ sparrow_cloud/access_control/decorators.py
@@ -3,8 +3,6 @@ import logging
 
 from functools import wraps
 
-from rest_framework.exceptions import PermissionDenied
-
 from django.http import HttpResponse
 from django.utils.decorators import method_decorator
 
@@ -13,7 +11,8 @@ from sparrow_cloud.utils.get_settings_value import get_settings_value
 
 logger = logging.getLogger(__name__)
 
-DETAIL = {"detail": "You do not have permission to perform this action."}
+DETAIL_403 = {"message": "无权限访问。"}
+DETAIL_401 = {"message": "身份认证信息未提供。"}
 
 
 def access_control_fbv(resource=None):
@@ -27,10 +26,10 @@ def access_control_fbv(resource=None):
             if skip_access_control is False or skip_access_control == 'False':
                 user_id = request.META["REMOTE_USER"]
                 if user_id is None:
-                    raise PermissionDenied()
+                    return HttpResponse(json.dumps(DETAIL_401), content_type='application/json; charset=utf-8', status=401)
                 app_name = get_settings_value("SERVICE_CONF").get("NAME", None)
                 if not access_verify(user_id=user_id, app_name=app_name, resource_code=resource):
-                    raise PermissionDenied()
+                    return HttpResponse(json.dumps(DETAIL_403), content_type='application/json; charset=utf-8', status=403)
                 return func(request, *args, **kwargs)
             return func(request, *args, **kwargs)
         return wrap
@@ -48,10 +47,10 @@ def access_control_cbv_all(resource=None):
                 if skip_access_control is False or skip_access_control == 'False':
                     user_id = request.META["REMOTE_USER"]
                     if user_id is None:
-                        return HttpResponse(json.dumps(DETAIL), content_type='application/json; charset=utf-8', status=403)
+                        return HttpResponse(json.dumps(DETAIL_401), content_type='application/json; charset=utf-8', status=401)
                     app_name = get_settings_value("SERVICE_CONF").get("NAME", None)
                     if not access_verify(user_id=user_id, app_name=app_name, resource_code=resource):
-                        return HttpResponse(json.dumps(DETAIL), content_type='application/json; charset=utf-8', status=403)
+                        return HttpResponse(json.dumps(DETAIL_403), content_type='application/json; charset=utf-8', status=403)
                     return function(request, *args, **kwargs)
                 return function(request, *args, **kwargs)
             return wrap
@@ -73,13 +72,13 @@ def access_control_cbv_method(resource):
                 skip_access_control = get_settings_value("ACCESS_CONTROL").get("SKIP_ACCESS_CONTROL", False)
                 if skip_access_control is False or skip_access_control == 'False':
                     user_id = request.META["REMOTE_USER"]
-                    if user_id is None:
-                        return HttpResponse(json.dumps(DETAIL), content_type='application/json; charset=utf-8', status=403)
+                    app_name = get_settings_value("SERVICE_CONF").get("NAME", None)
                     resource_code = (dict((k.lower(), v) for k, v in resource.items())).get(request.method.lower())
                     if resource_code:
-                        app_name = get_settings_value("SERVICE_CONF").get("NAME", None)
+                        if user_id is None:
+                            return HttpResponse(json.dumps(DETAIL_401), content_type='application/json; charset=utf-8', status=401)
                         if not access_verify(user_id=user_id, app_name=app_name, resource_code=resource_code):
-                            return HttpResponse(json.dumps(DETAIL), content_type='application/json; charset=utf-8', status=403)
+                            return HttpResponse(json.dumps(DETAIL_403), content_type='application/json; charset=utf-8', status=403)
                     return function(request, *args, **kwargs)
                 return function(request, *args, **kwargs)
             return wrap