+++ sparrow_cloud/apps/message_service/management/commands/_sparrow_rabbitmq_consumer.py
@@ -1,5 +1,5 @@
 from ._controller import RabbitMQConsumer
-from sparrow_cloud.registry.service_discovery import consul_service
+# from sparrow_cloud.registry.service_discovery import consul_service
 from sparrow_cloud.apps.message_service.aliyun_amqp import AliyunCredentialsProvider3 as aliyun_provider
 from django.conf import settings
 import time
@@ -27,10 +27,7 @@ def rabbitmq_consumer(queue):
                 "USER_NAME": "",
                 "PASSWORD": "",
                 "VIRTUAL_HOST": "",
-                "BROKER_SERVICE_CONF": {
-                    "ENV_NAME": "",
-                    "VALUE": "",
-                },
+                "BROKER_SERVICE_CONF": "sparrow-demo:8000",
             },
             "ALIYUN_RABBITMQ_BROKER": {
                 "HOST": "",
@@ -43,10 +40,7 @@ def rabbitmq_consumer(queue):
             }, 
             "RABBITMQ_SELECTION": "MESSAGE_BROKER_CONF",
             "MESSAGE_BACKEND_CONF": {
-                "BACKEND_SERVICE_CONF": {
-                        "ENV_NAME": "",
-                        "VALUE": "",
-                },
+                "BACKEND_SERVICE_CONF": "sparrow-demo:8000",
                 "API_PATH": ""
             },
             "RETRY_TIMES": 3,
@@ -109,10 +103,11 @@ def rabbitmq_consumer(queue):
                 username = provider.get_username()
                 password = provider.get_password()
             else:
-                broker_service_conf = rabbitmq_conf.get('BROKER_SERVICE_CONF', None)
+                # broker_service_conf = rabbitmq_conf.get('BROKER_SERVICE_CONF', None)
+                broker_service_addr = rabbitmq_conf.get('BROKER_SERVICE_CONF', None)
                 username = rabbitmq_conf.get('USER_NAME', None)
                 password = rabbitmq_conf.get('PASSWORD', None)
-                broker_service_addr = consul_service(broker_service_conf)
+                # broker_service_addr = consul_service(broker_service_conf)
                 host = broker_service_addr.split(':')[0]
                 port = broker_service_addr.split(':')[1]