+++ sparrow_cloud/dingtalk/sender.py
@@ -6,7 +6,7 @@ from sparrow_cloud.utils.get_settings_value import get_settings_value
 logger = logging.getLogger(__name__)
 
 
-def send_message(msg, code_list, channel="dingtalk", message_type="text"):
+def send_message(msg, code_list, channel="dingtalk", message_type="text", *args, **kwargs):
     """钉钉群发消息机器人 client """
     if not isinstance(msg, str) and not isinstance(code_list, list):
         raise TypeError("参数类型错误：msg type not string or code_list type not list")
@@ -18,7 +18,7 @@ def send_message(msg, code_list, channel="dingtalk", message_type="text"):
     }
     sparrow_ding_talk_conf = get_settings_value("SPARROW_DING_TALK_CONF")
     try:
-        res = rest_client.post(sparrow_ding_talk_conf["SERVICE_DING_TALK"], sparrow_ding_talk_conf["PATH"], data=data)
+        res = rest_client.post(sparrow_ding_talk_conf["SERVICE_DING_TALK"], sparrow_ding_talk_conf["PATH"], data=data, *args, **kwargs)
         logging.info("sparrow_cloud ding_talk send_message: msg:{}, group_code_list:{}".format(msg, code_list))
         return res
     except HTTPException as ex: