+++ tests/test_ding_talk.py
@@ -5,12 +5,12 @@ from sparrow_cloud.restclient.exception import HTTPException
 from django.test import RequestFactory
 from sparrow_cloud.dingtalk.sender import send_message
 
-CONSUL_RETURN_DATA = (
-    "name",
-    [
-        {'ServiceAddress': '162.23.7.247', 'ServicePort': 8001}
-    ]
-)
+# CONSUL_RETURN_DATA = (
+#     "name",
+#     [
+#         {'ServiceAddress': '162.23.7.247', 'ServicePort': 8001}
+#     ]
+# )
 
 MOCK_RESPONSE = {"data": {"code": 0, "message": "success"}}
 
@@ -53,10 +53,7 @@ class TestSendMessage(unittest.TestCase):
     def test_send_message(self, request):
         from django.conf import settings
         settings.SPARROW_DING_TALK_CONF = {
-            "SERVICE_DING_TALK": {
-                "ENV_NAME": "SERVICE_DING_TALK_HOST",
-                "VALUE": os.environ.get("SERVICE_DING_TALK", "ding-talk"),
-            },
+            "SERVICE_DING_TALK": "ding-talk:8001",
             "PATH": "/send/message/",
         }
         res = send_message("test", ["123"], "wechat", "text")
@@ -66,10 +63,7 @@ class TestSendMessage(unittest.TestCase):
     def test_send_message_400(self, request):
         from django.conf import settings
         settings.SPARROW_DING_TALK_CONF = {
-            "SERVICE_DING_TALK": {
-                "ENV_NAME": "SERVICE_DING_TALK_HOST",
-                "VALUE": os.environ.get("SERVICE_DING_TALK", "ding-talk"),
-            },
+            "SERVICE_DING_TALK": "ding-talk:8001",
             "PATH": "/send/message/",
         }
         self.assertEqual(send_message("test", ["123"], "wechat", "text").status_code, 400)
@@ -79,10 +73,7 @@ class TestSendMessage(unittest.TestCase):
     def test_send_message_typeerror(self, request):
         from django.conf import settings
         settings.SPARROW_DING_TALK_CONF = {
-            "SERVICE_DING_TALK": {
-                "ENV_NAME": "SERVICE_DING_TALK_HOST",
-                "VALUE": os.environ.get("SERVICE_DING_TALK", "ding-talk"),
-            },
+            "SERVICE_DING_TALK": "ding-talk:8001",
             "PATH": "/send/message/",
         }
         try: