+++ tests/test_message_client.py
@@ -10,10 +10,7 @@ class TestMessage(unittest.TestCase):
     def setUp(self):
         os.environ["SPARROW_TASK_TEST_SVC_HOST"] = "127.0.0.1:8001"
         settings.MESSAGE_SENDER_CONF = {
-            "SERVICE_CONF": {
-                "ENV_NAME": "SPARROW_TASK_TEST_SVC_HOST",
-                "VALUE": "sparrow-task-test-svc",
-            },
+            "SERVICE_CONF": "sparrow-task-test-svc:8001",
             "API_PATH": "/api/sparrow_task/producer/send/",
         }
 
@@ -27,3 +24,9 @@ class TestMessage(unittest.TestCase):
         self.assertEqual(data, {})
 
 
+    @mock.patch('sparrow_cloud.message_service.sender.TaskSender.send_task', return_value={})
+    def test_send_message_v2(self, mock_send_task):
+        from sparrow_cloud.message_service.sender import send_task_v2
+        message_code = 'order_pay_success'
+        data = send_task_v2(message_code=message_code)
+        self.assertEqual(data, {})