+++ tests/test_rabbitmq_consumer_command.py
@@ -1,9 +1,5 @@
-import unittest
-from unittest import mock
-import django
-from io import StringIO
-from django.core.management import call_command
 import os
+import unittest
 from django.conf.urls import url
 from django.http import HttpResponse
 
@@ -32,12 +28,12 @@ class RestClientTestCase(unittest.TestCase):
     # @mock.patch('rabbitmq_consumer.RabbitMQConsumer.target_func_map', return_value='')
     # @mock.patch('rabbitmq_consumer.RabbitMQConsumer.consume', return_value='接收任务成功')
     # def test_consumer_command(self, mock_target_func_map, mock_consume):
-        # from django.conf import settings
-        # self.setup_settings(settings)
-        # django.setup()
-        # out = StringIO()
-        # call_command('rabbitmq_consumer', '--queue', 'QUEUE_CONF', stdout=out)
-        # self.assertEqual(out.read(), '')
+    #     from django.conf import settings
+    #     self.setup_settings(settings)
+    #     django.setup()
+    #     out = StringIO()
+    #     call_command('rabbitmq_consumer', '--queue', 'QUEUE_CONF', stdout=out)
+    #     self.assertEqual(out.read(), '')
 
     def setup_settings(self, settings):
         settings.XX = "1"
@@ -47,16 +43,10 @@ class RestClientTestCase(unittest.TestCase):
                 "USER_NAME": "test_name",
                 "PASSWORD": "test_password",
                 "VIRTUAL_HOST": "test_virtual",
-                "BROKER_SERVICE_CONF": {
-                    "ENV_NAME": "SPARROW_BROKER_HOST",
-                    "VALUE": "sparrow-test",
-                },
+                "BROKER_SERVICE_CONF": "sparrow-test:8001",
             },
             "MESSAGE_BACKEND_CONF": {
-                "BACKEND_SERVICE_CONF": {
-                    "ENV_NAME": "SPARROW_BACKEND_HOST",
-                    "VALUE": "sparrow-demo",
-                },
+                "BACKEND_SERVICE_CONF": "sparrow-test:8001",
                 "API_PATH": "/api/sparrow_test/task/test_update/"
             }
         }